feat: centralize matchup types
