+++ lib/agents/lineWatcher.ts
@@ -1,9 +1,4 @@
-import { AgentResult } from './injuryScout';
-
-interface Matchup {
-  homeTeam: string;
-  awayTeam: string;
-}
+import { AgentResult, Matchup } from '../types';
 
 export const lineWatcher = (matchup: Matchup): AgentResult => {
   return {
